Repositories tests working.

+++ deliveries-engine/deliveries-engine/src/main/java/deliveries_engine/model/Rider.java
@@ -10,7 +10,7 @@ public class Rider {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @OneToOne
     @JoinColumn(name = "user_id", nullable = false)
@@ -27,7 +27,7 @@ public class Rider {
         this.status = false;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
